more loading stuff and misc cleanup
-clean up landcfg
-clean up some previously unnecessary changes
-change pos to vector2 and move z into own zOrder rendering variable due to error prone management of z axis.

+++ core/src/com/spaceproject/components/OrbitComponent.java
@@ -8,7 +8,7 @@ public class OrbitComponent implements Component {
 	public Entity parent; 
 	
 	//angle from parent in radians
-	public double angle, startAngle;
+	public float angle, startAngle;
 	
 	//orbit parameters
 	public float radialDistance; //distance from entity to rotate
